fix(core): default agent capabilities to scoped
